octofit-tracker/backend/requirements_minimal.txt
Django==4.1.7
djangorestframework==3.14.0
django-cors-headers==4.5.0
pymongo==3.12